Crate moxcms

Structs

Chromaticity

CicpProfile

A Coding Independent Code Point.

ColorDateTime

ColorPrimaries

ColorProfile

ICC Profile representation

Cube

3D CLUT helper

DescriptionString

DtUchHcb

Darktable HCB ( Darktable Uniform Color Space )

DtUchHsb

Darktable UCS HSB ( Darktable Uniform Color Space )

DtUchJch

Darktable UCS JCH ( Darktable Uniform Color Space )

Hypercube

4D CLUT helper.

ICtCp

Jzazbz

Represents Jzazbz

Jzczhz

Represents Jzazbz in polar coordinates as Jzczhz

LCh

Representing a color in cylindrical CIE LCh(uv) color space

Lab

Holds CIE LAB values

LocalizableString

LutDataType

LutMultidimensionalType

Luv

Struct representing a color in CIE LUV, a.k.a. L*u*v*, color space

MalformedSize

Matrix3

Matrix3d

Matrix math helper

Matrix3f

Matrix math helper

Matrix4f

Measurement

Oklab

Struct that represent Oklab colorspace

Oklch

Represents Oklch colorspace

ParsingOptions

Rgb

Represents any RGB values

Srlab2

TransformOptions

Declares additional transformation options

Vector3

Vector math helper

Vector4

Vector math helper

ViewingConditions

XyY

Holds CIE XyY representation

Xyz

Holds CIE XYZ representation

Xyzd

Holds CIE XYZ representation, in double precision

Ych

Structure for cone form of Yrg colorspace

Yrg

Structure for Yrg colorspace

Enums

BarycentricWeightScale

Barycentric interpolation weights size.

CicpColorPrimaries

See Rec. ITU-T H.273 (12/2016) Table 2 Values 0, 3, 13–21, 23–255 are all reserved so all map to the same variant

CmsError

DataColorSpace

InterpolationMethod

Defines the interpolation method.

Layout

Layout declares a data layout. For RGB it shows also the channel order. To handle different data bit-depth appropriate executor must be used. Cmyk8 uses the same layout as Rgba8.

LutStore

LutType

LutWarehouse

MatrixCoefficients

Matrix Coefficients Enum (from ISO/IEC 23091-4 / MPEG CICP)

MeasurementGeometry

ProfileClass

ProfileSignature

ProfileText

ProfileVersion

RenderingIntent

StandardIlluminant

StandardObserver

TechnologySignatures

ToneReprCurve

TransferCharacteristics

See Rec. ITU-T H.273 (12/2016) Table 3 Values 0, 3, 19–255 are all reserved so all map to the same variant

Constants

BT2020_MATRIX

DISPLAY_P3_MATRIX

HLG_LUT_TABLE

Hybrid Log Gamma Lookup table

PQ_LUT_TABLE

Perceptual Quantizer Lookup table

SRGB_MATRIX

WHITE_POINT_D50

WHITE_POINT_D60

WHITE_POINT_D65

WHITE_POINT_DCI_P3

Traits

FusedExp

FusedExp2

FusedExp10

FusedLog

FusedLog2

FusedLog10

FusedPow

InPlaceStage

Helper for intermediate transformation stages

Stage

Helper for intermediate transformation stages

ToneCurveEvaluator

TransformExecutor

Transformation executor itself

XyYRepresentable

Functions

adapt_to_d50

adapt_to_d50_d

adapt_to_illuminant

adapt_to_illuminant_d

adapt_to_illuminant_xyz

adapt_to_illuminant_xyz_d

adaption_matrix

adaption_matrix_d

cie_y_1931_to_cie_y_2006

curve_from_gamma

Creates Tone Reproduction curve from gamma

filmlike_clip

Soft clipping out-of-bounds values in S-curve

rounding_div_ceil

Type Aliases

Transform8BitExecutor

Transform16BitExecutor

TransformF32BitExecutor

TransformF64BitExecutor

Vector3d

Vector3f

Vector3i

Vector3u

Vector4d

Vector4f

Vector4i