moxcms::cicp

Enum TransferCharacteristics

pub enum TransferCharacteristics {

    Reserved = 0,
    Bt709 = 1,
    Unspecified = 2,
    Bt470M = 4,
    Bt470Bg = 5,
    Bt601 = 6,
    Smpte240 = 7,
    Linear = 8,
    Log100 = 9,
    Log100sqrt10 = 10,
    Iec61966 = 11,
    Bt1361 = 12,
    Srgb = 13,
    Bt202010bit = 14,
    Bt202012bit = 15,
    Smpte2084 = 16,
    Smpte428 = 17,
    Hlg = 18,

}

See Rec. ITU-T H.273 (12/2016) Table 3 Values 0, 3, 19–255 are all reserved so all map to the same variant

Variants

Reserved = 0

For future use by ITU-T | ISO/IEC

Bt709 = 1

Rec. ITU-R BT.709-6
Rec. ITU-R BT.1361-0 conventional colour gamut system (historical)
(functionally the same as the values 6, 14 and 15)

Unspecified = 2

Image characteristics are unknown or are determined by the application.

Bt470M = 4

Rec. ITU-R BT.470-6 System M (historical)
United States National Television System Committee 1953 Recommendation for transmission standards for color television
United States Federal Communications Commission (2003) Title 47 Code of Federal Regulations 73.682 (a) (20)
Rec. ITU-R BT.1700-0 625 PAL and 625 SECAM

Bt470Bg = 5

Rec. ITU-R BT.470-6 System B, G (historical)

Bt601 = 6

Rec. ITU-R BT.601-7 525 or 625
Rec. ITU-R BT.1358-1 525 or 625 (historical)
Rec. ITU-R BT.1700-0 NTSC SMPTE 170M (2004)
(functionally the same as the values 1, 14 and 15)

Smpte240 = 7

SMPTE 240M (1999) (historical)

Linear = 8

Linear transfer characteristics

Log100 = 9

Logarithmic transfer characteristic (100:1 range)

Log100sqrt10 = 10

Logarithmic transfer characteristic (100 * Sqrt( 10 ) : 1 range)

Iec61966 = 11

IEC 61966-2-4

Bt1361 = 12

Rec. ITU-R BT.1361-0 extended colour gamut system (historical)

Srgb = 13

IEC 61966-2-1 sRGB or sYCC

Bt202010bit = 14

Rec. ITU-R BT.2020-2 (10-bit system)
(functionally the same as the values 1, 6 and 15)

Bt202012bit = 15

Rec. ITU-R BT.2020-2 (12-bit system)
(functionally the same as the values 1, 6 and 14)

Smpte2084 = 16

SMPTE ST 2084 for 10-, 12-, 14- and 16-bitsystems
Rec. ITU-R BT.2100-0 perceptual quantization (PQ) system

Smpte428 = 17

SMPTE ST 428-1

Hlg = 18

ARIB STD-B67
Rec. ITU-R BT.2100-0 hybrid log- gamma (HLG) system

Implementations

impl TransferCharacteristics

pub(crate) fn has_transfer_curve(self) -> bool

impl TransferCharacteristics

pub fn linearize(self, v: f64) -> f64

pub fn gamma(self, v: f64) -> f64

pub(crate) fn make_linear_table<T: PointeeSizeExpressible, const N: usize, const BIT_DEPTH: usize>( &self, ) -> Box<[f32; N]>

pub(crate) fn make_gamma_table<T: Default + Copy + 'static + PointeeSizeExpressible, const BUCKET: usize, const N: usize>( &self, bit_depth: usize, ) -> Box<[T; BUCKET]>
where f32: AsPrimitive<T>,