moxcms::conversions

Module rgbxyz

Macros

create_in_place_opt_rgb_xyz_fp_to_v 🔒

create_in_place_rgb_xyz 🔒

create_rgb_xyz_dependant_executor 🔒

create_rgb_xyz_dependant_executor_to_v 🔒

Structs

TransformMatrixShaper 🔒

TransformMatrixShaperOptScalar 🔒

TransformMatrixShaperOptimized 🔒

Low memory footprint optimized routine for matrix shaper profiles with the same Gamma and linear curves.

TransformMatrixShaperOptimizedV 🔒

Low memory footprint optimized routine for matrix shaper profiles with the same Gamma and linear curves.

TransformMatrixShaperScalar 🔒

TransformMatrixShaperV 🔒

Functions

make_rgb_xyz_rgb_transform 🔒

make_rgb_xyz_rgb_transform_avx2 🔒

make_rgb_xyz_rgb_transform_opt 🔒

make_rgb_xyz_rgb_transform_sse_41 🔒