moxcms

Enum CicpColorPrimaries

pub enum CicpColorPrimaries {

    Reserved = 0,
    Bt709 = 1,
    Unspecified = 2,
    Bt470M = 4,
    Bt470Bg = 5,
    Bt601 = 6,
    Smpte240 = 7,
    GenericFilm = 8,
    Bt2020 = 9,
    Xyz = 10,
    Smpte431 = 11,
    Smpte432 = 12,
    Ebu3213 = 22,

}

See Rec. ITU-T H.273 (12/2016) Table 2 Values 0, 3, 13โ21, 23โ255 are all reserved so all map to the same variant

Variants

Reserved = 0

For future use by ITU-T | ISO/IEC

Bt709 = 1

Rec. ITU-R BT.709-6
Rec. ITU-R BT.1361-0 conventional colour gamut system and extended colour gamut system (historical)
IEC 61966-2-1 sRGB or sYCC IEC 61966-2-4
Society of Motion Picture and Television Engineers (MPTE) RP 177 (1993) Annex B

Unspecified = 2

Unspecified
Image characteristics are unknown or are determined by the application.

Bt470M = 4

Rec. ITU-R BT.470-6 System M (historical)
United States National Television System Committee 1953 Recommendation for transmission standards for color television
United States Federal Communications Commission (2003) Title 47 Code of Federal Regulations 73.682 (a) (20)

Bt470Bg = 5

Rec. ITU-R BT.470-6 System B, G (historical) Rec. ITU-R BT.601-7 625
Rec. ITU-R BT.1358-0 625 (historical)
Rec. ITU-R BT.1700-0 625 PAL and 625 SECAM

Bt601 = 6

Rec. ITU-R BT.601-7 525
Rec. ITU-R BT.1358-1 525 or 625 (historical) Rec. ITU-R BT.1700-0 NTSC
SMPTE 170M (2004)
(functionally the same as the value 7)

Smpte240 = 7

SMPTE 240M (1999) (historical) (functionally the same as the value 6)

GenericFilm = 8

Generic film (colour filters using Illuminant C)

Bt2020 = 9

Rec. ITU-R BT.2020-2
Rec. ITU-R BT.2100-0

Xyz = 10

SMPTE ST 428-1
(CIE 1931 XYZ as in ISO 11664-1)

Smpte431 = 11

SMPTE RP 431-2 (2011)

Smpte432 = 12

SMPTE EG 432-1 (2010)

Ebu3213 = 22

EBU Tech. 3213-E (1975)