moxcms

Struct LCh

#[repr(C)]
pub struct LCh {
    pub l: f32,
    pub c: f32,
    pub h: f32,
}

Representing a color in cylindrical CIE LCh(uv) color space

Fields
l: f32

The L* value (achromatic luminance) of the colour in 0–100 range.

This is the same value as in the Luv object.

c: f32

The C*_uv value (chroma) of the colour.

Together with h_uv, it defines chromaticity of the colour. The typical values of the coordinate go from zero up to around 150 (but exact range for ‘valid’ colours depends on luminance and hue). Zero represents shade of grey.

h: f32

The h_uv value (hue) of the colour measured in radians.

Together with C*_uv, it defines chromaticity of the colour. The value represents an angle thus it wraps around τ. Typically, the value will be in the -π–π range. The value is undefined if C*_uv is zero.

Implementations

impl LCh

pub const fn new(l: f32, c: f32, h: f32) -> Self

pub fn from_luv(luv: Luv) -> Self

Converts Lab to LCh(uv)

pub fn from_lab(lab: Lab) -> Self

Converts Lab to LCh(ab)

pub fn from_xyz(xyz: Xyz) -> Self

Computes LCh(uv)

pub fn from_xyz_lab(xyz: Xyz) -> Self

Computes LCh(ab)

pub fn to_xyz(&self) -> Xyz

Converts LCh(uv) to Luv

pub fn to_xyz_lab(&self) -> Xyz

Converts LCh(ab) to Lab

pub fn to_luv(&self) -> Luv

pub fn to_lab(&self) -> Lab

impl LCh

pub fn euclidean_distance(&self, other: LCh) -> f32

impl LCh

pub const fn taxicab_distance(&self, other: Self) -> f32

impl LCh

pub fn sqrt(&self) -> LCh

pub fn cbrt(&self) -> LCh