moxcms

Struct Rgb

#[repr(C)]
pub struct Rgb<T> {
    pub r: T,
    pub g: T,
    pub b: T,
}

Represents any RGB values

Fields
r: T

Red component

g: T

Green component

b: T

Blue component

Implementations

impl<T> Rgb<T>

pub fn new(r: T, g: T, b: T) -> Rgb<T>

impl<T> Rgb<T>
where T: Copy,

pub fn dup(v: T) -> Rgb<T>

pub const fn to_vector(self) -> Vector3<T>

impl Rgb<f32>

pub fn apply(&self, matrix: Matrix3f) -> Rgb<f32>

pub fn to_xyz(&self, matrix: Matrix3f) -> Xyz

pub fn is_out_of_gamut(&self) -> bool

impl Rgb<f32>

pub fn zeroed() -> Rgb<f32>

pub fn ones() -> Rgb<f32>

pub fn white() -> Rgb<f32>

pub fn black() -> Rgb<f32>

impl Rgb<f64>

pub fn zeroed() -> Rgb<f64>

pub fn ones() -> Rgb<f64>

pub fn white() -> Rgb<f64>

pub fn black() -> Rgb<f64>

impl Rgb<u8>

pub fn zeroed() -> Rgb<u8>

pub fn capped() -> Rgb<u8>

pub fn white() -> Rgb<u8>

pub fn black() -> Rgb<u8>

impl Rgb<u16>

pub fn zeroed() -> Rgb<u16>

pub fn capped() -> Rgb<u16>

pub fn white() -> Rgb<u16>

pub fn black() -> Rgb<u16>

impl Rgb<i8>

pub fn zeroed() -> Rgb<i8>

pub fn capped() -> Rgb<i8>

pub fn white() -> Rgb<i8>

pub fn black() -> Rgb<i8>

impl Rgb<i16>

pub fn zeroed() -> Rgb<i16>

pub fn capped() -> Rgb<i16>

pub fn white() -> Rgb<i16>

pub fn black() -> Rgb<i16>

impl Rgb<i32>

pub fn zeroed() -> Rgb<i32>

pub fn capped() -> Rgb<i32>

pub fn white() -> Rgb<i32>

pub fn black() -> Rgb<i32>

impl Rgb<u32>

pub fn zeroed() -> Rgb<u32>

pub fn capped() -> Rgb<u32>

pub fn white() -> Rgb<u32>

pub fn black() -> Rgb<u32>

impl<T> Rgb<T>
where T: Copy + AsPrimitive<f32>,

pub fn euclidean_distance(&self, other: Rgb<T>) -> f32

impl<T> Rgb<T>
where T: Copy + AsPrimitive<f32>,

pub fn taxicab_distance(&self, other: Self) -> f32

impl<T> Rgb<T>
where T: Signed + Copy + Clone,

pub fn abs(self) -> Self

impl<T> Rgb<T>
where T: FusedMultiplyAdd<T>,

pub fn mla(&self, b: Rgb<T>, c: Rgb<T>) -> Rgb<T>

impl<T> Rgb<T>
where T: Num + PartialOrd + Copy + Bounded,

pub fn clamp(&self, min_value: T, max_value: T) -> Rgb<T>

Clamp function to clamp each channel within a given range

pub fn min(&self, other_min: T) -> Rgb<T>

Min function to define min

pub fn max(&self, other_max: T) -> Rgb<T>

Max function to define max

pub fn clamp_p(&self, min_value: Rgb<T>, max_value: Rgb<T>) -> Rgb<T>

Clamp function to clamp each channel within a given range

pub fn min_p(&self, other_min: Rgb<T>) -> Rgb<T>

Min function to define min

pub fn max_p(&self, other_max: Rgb<T>) -> Rgb<T>

Max function to define max

impl<T> Rgb<T>
where T: Float + 'static, f32: AsPrimitive<T>,

pub fn sqrt(&self) -> Rgb<T>

pub fn cbrt(&self) -> Rgb<T>

impl<T> Rgb<T>

pub fn cast<V>(self) -> Rgb<V>
where T: AsPrimitive<V>, V: Copy + 'static,

impl<T> Rgb<T>
where T: Float + 'static,

pub fn round(self) -> Rgb<T>