moxcms::transform

Trait TransformExecutor

pub trait TransformExecutor<V: Copy + Default> {
    // Required method
    fn transform(&self, src: &[V], dst: &mut [V]) -> Result<(), CmsError>;
}

Transformation executor itself

Required Methods

fn transform(&self, src: &[V], dst: &mut [V]) -> Result<(), CmsError>

Count of samples always must match. If there is N samples of Cmyk source then N samples of Rgb is expected as an output.